Added long term 'down' detection to the scanner
Added [scan]->down to the config file to limit scanning of games that appear to
have gone permanantly offline.  Detection is overidden by a successful manual
connection.

Bumped db version to accomodate a new field in the scan table.  Normally I like
the db version number to be a YYMMDD encoding, but I guess I messed that up the
previous time that I bumped the db number.  Intent is to go back to the YYMMDD
format in 2026.