Add viewport interactive-widget resizes-content
Add interactive-widget=resizes-content to the viewport to re-enable the 'old'
android behavior where opening the virtualKeyboard would trigger a full resize
of the lociterm app.  Has the downside of sending NAWS every time the keyboard
pops open, but the upside of keeping the top anchored menubox on screen, and
any any game-side pager aware of the line height that is actually on screen.

Also add Ctrl-M to the menuhandlers as additional shortcut for close menu.
Intent is to convert into a focus toggle (instead of a close) in a future
commit.