Added meta_build information to versioning.
See makefile.  Changes LOCITERM_VERSION to be built from a LOCITERM_CORE
semver, and an prerelease tag based on the local codebase version in META_NAME
and META_VERSION.  This is to better support version control for sites that
want to maintain game specific changes changes that are not part of upstream
lociterm.